v0.1.1
- Refactored MCP tools and resources into separate files

- Finished createSquareInvoice (Needs more testing)

- Removed database.ts

-Renamed debug_tools to database_debug_tools
