+++ package.json
@@ -23,7 +23,7 @@
     "dotenv": "^17.2.1",
     "libphonenumber-js": "^1.12.15",
     "square": "^43.0.2",
-    "zod": "^4.1.4"
+    "zod": "^3.25.76"
   },
   "devDependencies": {
     "@types/node": "^20.19.11",
