+++ src/clients/prismaClient.ts
@@ -1,0 +1,6 @@
+import { PrismaClient } from "@prisma/client";
+
+// Init Prisma Client
+const prisma = new PrismaClient();
+
+export default prisma;
